Move sandbox isolation code to a new file
